pyproject.toml
[project]
name = "romanhistory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bm25s>=0.2.14",
    "flagembedding>=1.3.5",
    "gradio>=5.49.1",
    "langgraph>=0.6.7",
    "pgvector>=0.4.2",
    "psycopg>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pystemmer>=3.0.0",
    "sentence-transformers>=3.0.0",
    "together>=1.5.25",
    "transformers>=4.51,<5",
]
